refactor: remove cloud-status feature and integrate status management
- Deleted the cloud-status feature, including its translations, model, and UI components, to streamline the application.
- Updated the i18n configuration to incorporate status management instead of cloud-status.
- Refactored related imports and tests to ensure consistency with the new status management approach.
- Enhanced the LoginForm e2e tests for improved clarity and reliability.
