--- src/renderer/features/cloud-status/i18n/keys.ts
@@ -1,6 +1,0 @@
-const translationKeys = {
-  cloud: 'cloudStatus.cloud',
-  cloudless: 'cloudStatus.cloudless'
-}
-
-export default translationKeys
