--- src/renderer/features/network-status/i18n/keys.ts
@@ -1,8 +1,0 @@
-const translationKeys = {
-  online: 'networkStatus.online',
-  offline: 'networkStatus.offline',
-  cloud: 'networkStatus.cloud',
-  cloudless: 'networkStatus.cloudless'
-}
-
-export default translationKeys
