+++ src/renderer/features/status/cloud-status/i18n/en.ts
