+++ src/renderer/features/status/cloud-status/i18n/index.ts
@@ -1,0 +1,2 @@
+export { default as cloudStatusDe } from './de'
+export { default as cloudStatusEn } from './en'
