+++ src/renderer/features/status/cloud-status/i18n/keys.ts
@@ -1,0 +1,6 @@
+const translationKeys = {
+  cloud: 'status.cloud.cloud',
+  cloudless: 'status.cloud.cloudless'
+}
+
+export default translationKeys
