+++ src/renderer/features/status/cloud-status/service/cloud-status-storage.test.ts
