+++ src/renderer/features/status/cloud-status/store/index.ts
