+++ src/renderer/features/status/cloud-status/ui/CloudStatus.e2e.spec.ts
