+++ src/renderer/features/status/cloud-status/ui/CloudStatus.vue
