+++ src/renderer/features/status/i18n/de.ts
@@ -1,0 +1,7 @@
+import { cloudStatusDe } from '../cloud-status/i18n'
+import { networkStatusDe } from '../network-status/i18n'
+
+export default {
+  cloud: cloudStatusDe,
+  network: networkStatusDe
+}
