+++ src/renderer/features/status/i18n/en.ts
@@ -1,0 +1,7 @@
+import { cloudStatusEn } from '../cloud-status/i18n'
+import { networkStatusEn } from '../network-status/i18n'
+
+export default {
+  cloud: cloudStatusEn,
+  network: networkStatusEn
+}
