+++ src/renderer/features/status/network-status/i18n/de.ts
