+++ src/renderer/features/status/network-status/i18n/en.ts
