+++ src/renderer/features/status/network-status/i18n/index.ts
@@ -1,0 +1,2 @@
+export { default as networkStatusDe } from './de'
+export { default as networkStatusEn } from './en'
