+++ src/renderer/features/status/network-status/i18n/keys.ts
@@ -1,0 +1,6 @@
+const translationKeys = {
+  online: 'status.network.online',
+  offline: 'status.network.offline'
+}
+
+export default translationKeys
