+++ src/renderer/features/status/network-status/model/use-network-status.ts
