+++ src/renderer/features/status/network-status/ui/NetworkStatus.e2e.spec.ts
