+++ src/renderer/features/status/network-status/ui/NetworkStatus.vue
