+++ src/renderer/shared/lib/index.ts
@@ -13,4 +13,4 @@ export { useTranslation } from './i18n/use-translation'
 export { getActiveStore, newStore, newStoreToRefs } from './pinia/store-define'
 export { ErrorCode, errorTranslationMap, translateError } from './validation/error-manager'
 export { emailRules, mapRule, PASSWORD_MIN_LENGTH, passwordRules } from './validation/validators'
-export { useCloudStatusStore } from '@features/cloud-status'
+export { useCloudStatusStore } from '@features/status'
