+++ src/renderer/shared/store/network/index.ts
@@ -1,2 +1,2 @@
 export { useNetworkStatusStore } from './use-network-status-store'
-export { useCloudStatusStore } from '@features/cloud-status'
+export { useCloudStatusStore } from '@features/status'
